build metacat.properties dynamically
Closes #2

The current solution is to use a python script which applies
the configuration in the `docker-entrypoint.sh` on startup.
The script will exit in error if any properties set in the
local application properties does not exist in the default
metacat.properties.

In the future, it is understood that the metacat admin utility
will perform this merge on startup.