Refactor build scripts to allow building local and registry images
Issue ess-dive/essdive-appstack#12

When building images for a local repository the registry prefix (e.g
registry.spin.nersc.gov/m2690) to an image is unnecessary. This will
require a logic change to all docker image repository build.sh scripts.

The change is to ignore REGISTRY_SPIN if it is unset